requirements.txt
beautifulsoup4
python-dotenv
requests
difflib
urllib3
openai